pyproject.toml
[project]
name = "x-feeder"
version = "0.1.0"
description = ""
authors = [
    {name = "k-kan-dev",email = "k.kan.splended@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "supabase (>=2.13.0,<3.0.0)",
    "discord (>=2.3.2,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "rich (>=13.9.4,<14.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
